requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
asgiref==3.7.2
    # via django
build==1.0.3
    # via pip-tools
click==8.1.7
    # via pip-tools
crispy-bootstrap5==2023.10
    # via -r requirements.in
django==5.0.2
    # via
    #   -r requirements.in
    #   crispy-bootstrap5
    #   django-crispy-forms
    #   django-localflavor
django-crispy-forms==2.1
    # via
    #   -r requirements.in
    #   crispy-bootstrap5
django-environ==0.11.2
    # via -r requirements.in
django-localflavor==4.0
    # via -r requirements.in
packaging==23.2
    # via build
pip-tools==7.3.0
    # via -r requirements.in
pyproject-hooks==1.0.0
    # via build
python-stdnum==1.19
    # via django-localflavor
sqlparse==0.4.4
    # via django
tomli==2.0.1
    # via
    #   build
    #   pip-tools
    #   pyproject-hooks
typing-extensions==4.9.0
    # via asgiref
wheel==0.42.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
